requirements.txt
Pillow==7.0.0
psutil==5.6.7
pygame==1.9.6
python-vlc==3.0.7110
